factored out the django-independent cas functionality
